requirements.txt
streamlit>=1.24.0
pandas>=2.0.3
numpy>=1.24.3
python-fuzzywuzzy>=0.18.0
plotly>=5.15.0
openpyxl>=3.1.2
scipy>=1.11.1
python-Levenshtein>=0.21.1